requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-jose==3.3.0
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
